app/backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
opencv-python-headless
